Fix build: cross-platform key injection, Arduino 3.x type mismatches
- Move ESP_INSIGHT_AUTH_KEY injection to version.py script to avoid
  PlatformIO $ENV{} Windows expansion bug (expanded to entire os.environ)
- Fix max() type mismatch in WebUI.cpp (uint32_t vs unsigned int, GCC 13)

+++ src/WebUI.cpp
@@ -1710,7 +1710,7 @@ void WebUI::handleScreen() {
         return;
     }
 
-    uint32_t refreshSec = max(5u, _cfg.pollIntervalMs / 1000);
+    uint32_t refreshSec = max((uint32_t)5, _cfg.pollIntervalMs / 1000u);
     String html =
         "<!DOCTYPE html><html><head>"
         "<title>PlaneTracker - Screen</title>"
